Backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==46.0.6
Django==5.2.11
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mssql-django==1.6
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.1.1
prompt_toolkit==3.0.52
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.1.0
pyodbc==5.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
twilio==9.6.1
tzdata==2026.1
tzlocal==5.3.1
ua-parser==1.0.2
ua-parser-builtins==202603
uritemplate==4.2.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.6.0
